rename responses.py -> content.py
